userProfile and moods added

+++ users/views.py
@@ -1,8 +1,11 @@
-
-from django.contrib.auth.models import User, Group
+from django.contrib.auth.models import User
+from rest_framework.authtoken.models import Token
 from django.http import JsonResponse
 from rest_framework import viewsets
-from users.serializers import UserSerializer
+from rest_framework.response import Response
+from rest_framework.decorators import api_view
+from users.serializers import UserSerializer, MoodSerializer
+from users.models import Mood,UserProfile
 
 
 class UserViewSet(viewsets.ModelViewSet):
@@ -12,61 +15,61 @@ class UserViewSet(viewsets.ModelViewSet):
     queryset = User.objects.all().order_by('-date_joined')
     serializer_class = UserSerializer
 
+@api_view(['GET'])
+def getUsersMoods(request):
+    #token = Token.objects.get(key=request.GET.get('access_token'))
+    #print token
+    #user = token.user
+    userProfile = UserProfile.objects.get(user=request.user)
+    moods = Mood.objects.filter(userProfile=userProfile).order_by('-created')
 
+    print moods
 
+    serializer = MoodSerializer(moods, many=True)
+    return Response(serializer.data)
 
-from rest_framework.authtoken.models import Token
-
-def token_request(request):
-        print request
-        new_token = Token.objects.create(user=request.user)
-        return new_token
 
 
 from oauthlib.common import generate_token
-from django.contrib.auth import login
-
 from social.apps.django_app.utils import psa
-from django.contrib.auth.models import User
+from helpers import facebookHelper
+import models
+
+
 
 @psa('social:complete')
 def register_by_access_token(request, backend):
     # This view expects an access_token GET parameter, if it's needed,
     # request.backend and request.strategy will be loaded with the current
     # backend and strategy.
-    third_party_token = request.GET.get('access_token')
-
-    print third_party_token
 
 
-
-    user = User.objects.get(username='bartoszcwynar')
-
-    #user = request.backend.do_auth(third_party_token)
-
-    if user:
-        #login(request, user)
-
-        # We get our app!
-        #app = Application.objects.get(name="myapp")
-
-        # We delete the old token
+    if facebookHelper.validateUserByToken(request.GET.get('access_token')):
         try:
-            old = Token.objects.get(user=user)
+            user  = User.objects.get(username=request.GET.get('USER'))
         except:
-            pass
+            user = User.objects.create_user(request.GET.get('USER'))
+            models.createUserProfile(user)
+            print
+        if user:
+            try:
+                old = Token.objects.get(user=user)
+            except:
+                pass
+            else:
+                old.delete()
+
+            new_token = generate_token()
+
+            Token.objects.create(user=user,
+                                key=new_token)
+
+            print "------------------"
+            print new_token
+            print "-------------------"
+            return JsonResponse({"token": new_token})
         else:
-            old.delete()
+            return "ERROR"
 
-        # We create a new one
-
-        my_token = generate_token()
-        # We create the access token
-        # (we could create a refresh token too the same way)
-        Token.objects.create(user=user,
-                            key=my_token)
-        print my_token
-
-        return JsonResponse({"token": my_token})
     else:
-        return "ERROR"
+        return JsonResponse({"lipa": "whuj"})
